Fixing app performance attempt
