+++ .expo/packager-info.json
@@ -1,10 +1,10 @@
 {
   "expoServerPort": 19000,
   "packagerPort": 19001,
-  "packagerPid": 7928,
+  "packagerPid": 17632,
   "expoServerNgrokUrl": "https://fn-29n.minhchinh97h.tower-client.exp.direct",
   "packagerNgrokUrl": "https://packager.fn-29n.minhchinh97h.tower-client.exp.direct",
-  "ngrokPid": 1780,
+  "ngrokPid": 7852,
   "devToolsPort": 19002,
   "webpackServerPort": null
 }
