+++ components/main/bottomTabNavigator/layouts/modal-component/calendar/day-annotation/DayAnnotationPanel.Container.js
@@ -1,0 +1,12 @@
+import {connect} from 'react-redux'
+
+import DayAnnotationPanel from './DayAnnotationPanel.Presentational'
+
+const mapStateToProps = (state) => ({
+    monthComponent_arr: state.monthComponent_arr
+})
+
+export default connect(
+    mapStateToProps,
+    null
+)(DayAnnotationPanel)
