+++ components/main/header/Header.js
@@ -18,6 +18,10 @@ const styles = StyleSheet.create({
   },
 });
 
+import DayAnnotationCalendarFirstInitContainer from '../../../first-init/calendar/day-annotation-calendar/DayAnnotationCalendarFirstInit.Container'
+
+let MonthComponent_arr = []
+
 export default class Header extends React.Component {
 
     state = {
@@ -27,10 +31,14 @@ export default class Header extends React.Component {
   
 
     componentDidMount = () => {
+
     }
     
     render() {
       return (
+        <>
+        
+        <DayAnnotationCalendarFirstInitContainer />
 
         <View style={{
             paddingTop: 20,
@@ -70,6 +78,8 @@ export default class Header extends React.Component {
 
             
         </View>
+
+        </>
       );
     }
 }
