+++ first-init/calendar/day-annotation-calendar/actions/updateMonthComponentArr.js
@@ -1,0 +1,4 @@
+export const updateMonthComponentArr = monthComponent_arr => ({
+    type: 'UPDATE_DAY_ANNOTATION_MONTH_COMPONENT_ARR',
+    monthComponent_arr
+})
