+++ reducers/index.js
@@ -2,11 +2,13 @@ import {combineReducers} from 'redux'
 import {allTodos} from './allTodos'
 import {allCompletedTodos} from './allCompletedTodos'
 import {currentRoute} from './currentRoute'
+import {monthComponent_arr} from './calendar/day-annotation-calendar/dayAnnotationCalendar'
 
 const rootReducer = combineReducers({
     allTodos,
     allCompletedTodos,
-    currentRoute
+    currentRoute,
+    monthComponent_arr
 })
 
 export default rootReducer
